requirements.txt
disnake==2.9.1
aiosqlite==0.19.0
PyYAML~=6.0.1
python-dotenv~=1.0.0